requirements.txt
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
feedparser==6.0.11
Flask==3.1.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
requests==2.32.4
sgmllib3k==1.0.0
six==1.17.0
SQLAlchemy==2.0.41
typeguard==4.4.4
typing_extensions==4.14.0
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
xmltodict==0.14.2
